test(modules): shared version-aware InMemorySource
